+++ packages/react-widgets/src/widgets/TableWidget.js
@@ -14,7 +14,7 @@ import { _FeatureFlags, _hasFeatureFlag } from '@carto/react-core';
  * @param  {string} props.title - Title to show in the widget header.
  * @param  {string} props.dataSource - ID of the data source to get the data from.
  * @param  {Column[]} props.columns - List of data columns to display.
- * @param  {Column[]} props.hiddenColumnsFields - List of data columns to be retrieved, but not displayed.
+ * @param  {Column[]} props.hiddenColumnFields - List of data columns to be retrieved, but not displayed.
  * @param  {Function=} props.onRowClick - Function to handle on click events on rows.
  * @param  {Function=} [props.onError] - Function to handle error messages from the widget.
  * @param  {Function=} [props.onStateChange] - Callback to handle state updates of widgets
@@ -34,7 +34,7 @@ function TableWidget({
   title,
   dataSource,
   columns,
-  hiddenColumnsFields = [],
+  hiddenColumnFields = [],
   wrapperProps,
   noDataAlertProps,
   onError,
@@ -64,7 +64,7 @@ function TableWidget({
     id,
     dataSource,
     params: {
-      columns: [...columns.map((c) => c.field), ...hiddenColumnsFields],
+      columns: [...columns.map((c) => c.field), ...hiddenColumnFields],
       sortBy,
       sortDirection,
       sortByColumnType,
@@ -144,7 +144,7 @@ TableWidget.propTypes = {
       align: PropTypes.oneOf(['left', 'right'])
     })
   ).isRequired,
-  hiddenColumnsFields: PropTypes.arrayOf(PropTypes.string),
+  hiddenColumnFields: PropTypes.arrayOf(PropTypes.string),
   onError: PropTypes.func,
   wrapperProps: PropTypes.object,
   noDataAlertProps: PropTypes.object,